+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file. This projec
 - Can now assert that the `data` member of a document is `null`.
 - Can now assert that a resource object matches an expected structure.
 - Can now assert that the `data` member of a document is a resource identifier.
+- Can now assert that a document does not contain an `errors` member.
 
 ### Changed
 - Added a generic JSON API object tester class containing common assertions.